Enhance API of charts
