+++ src/components/expanded-state/chart/chart-data-labels/useRatio.js
@@ -3,7 +3,7 @@ import { useChartData } from 'react-native-animated-charts';
 import useReactiveSharedValue from 'react-native-animated-charts/helpers/useReactiveSharedValue';
 
 export function useRatio(name) {
-  const { nativeY, data } = useChartData();
+  const { originalY, data } = useChartData();
 
   const firstValue = useReactiveSharedValue(
     data?.points?.[0]?.y,
@@ -17,7 +17,7 @@ export function useRatio(name) {
   return useDerivedValue(
     () =>
       firstValue.value === Number(firstValue.value)
-        ? (nativeY.value || lastValue.value) / firstValue.value
+        ? (originalY.value || lastValue.value) / firstValue.value
         : 1,
     undefined,
     name ? 'ratio_' + name : undefined
