+++ src/react-native-animated-charts/.release-it.json
@@ -1,0 +1,8 @@
+{
+    "git": {
+      "commitMessage": "chore: release v${version}"
+    },
+    "github": {
+      "release": true
+    }
+  }
