+++ src/react-native-animated-charts/package.json
@@ -12,11 +12,13 @@
   ],
   "license": "ISC",
   "dependencies": {
-    "@babel/core": "^7.11.4",
-    "@babel/runtime": "^7.11.2",
     "react-native-gesture-handler": "^1.7.0",
     "react-native-reanimated": "^2.0.0-alpha.5",
     "react-native-svg": "^12.1.0"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.11.4",
+    "@babel/runtime": "^7.11.2"
   },
   "peerDependencies": {
     "react": "*",
