+++ src/react-native-animated-charts/src/charts/linear/ChartDot.js
@@ -1,11 +1,24 @@
 import React, { useContext } from 'react';
 // eslint-disable-next-line import/no-unresolved
-import Animated from 'react-native-reanimated';
+import Animated, { useAnimatedStyle } from 'react-native-reanimated';
 import ChartContext from '../../helpers/ChartContext';
 import withReanimatedFallback from '../../helpers/withReanimatedFallback';
 
 function ChartDot({ style, size = 10, ...props }) {
-  const { dotStyle } = useContext(ChartContext);
+  const { dotScale, positionX, positionY } = useContext(ChartContext);
+  const dotStyle = useAnimatedStyle(
+    () => ({
+      opacity: dotScale.value,
+      transform: [
+        { translateX: positionX.value },
+        { translateY: positionY.value + 10 }, // TODO temporary fix for clipped chart
+        { scale: dotScale.value },
+      ],
+    }),
+    undefined,
+    'dotStyle'
+  );
+
   return (
     <Animated.View
       {...props}
