+++ src/react-native-animated-charts/src/helpers/ChartContext2.js
@@ -1,0 +1,2 @@
+import { createContext } from 'react';
+export default createContext(null);
