+++ src/react-native-animated-charts/src/helpers/useChartData.js
@@ -2,6 +2,5 @@ import { useContext } from 'react';
 import ChartContext from './ChartContext';
 
 export default function useChartData() {
-  const { extremes, ...rest } = useContext(ChartContext);
-  return { ...extremes, ...rest };
+  return useContext(ChartContext);
 }
